* added AudioRecorderWrapper device driver * deprecated portaudio device driver * deprecated ServerSoundGrabber device driver * improved methods IAudioGrabberSound::getSound() * improved devices PortAudioRecorderDeviceDriver, PortAudioPlayerDeviceDriver, AudioPlayerWrapper, AudioRecorderWrapper

+++ src/devices/portaudio/PortAudioBuffer.h
@@ -130,8 +130,8 @@ struct circularDataBuffers
     yarp::dev::CircularAudioBuffer_16t*     recData;
     bool                canPlay;
     bool                canRec;
-    int                 numPlayChannels;
-    int                 numRecChannels;
+    size_t              numPlayChannels;
+    size_t              numRecChannels;
     circularDataBuffers ()
     {
         numPlayChannels = 1;
